Don't pollute the main repo directory
GitHub CI jobs run in the `${GITHUB_WORKSPACE}` directory, which is of the form
`.../repo-name/repo-name`. The checkout@v3 action then replaces the contents of
that directory with the checked out repository. Rather than polluting that same
directory with the opendylan installation, which could cause conflicts when its
projects are added to the registry generated by `dylan update`, we now install
Open Dylan in `.../repo-name/opendylan` instead.